+++ src/main/scala/cromwell/engine/backend/package.scala
@@ -1,8 +1,14 @@
 package cromwell.engine
 
 import scala.concurrent.{ExecutionContext, Future}
+import scala.collection.immutable.Seq
 
 package object backend {
+  // Ordered by shards, and then ordered by attempts
+  type AttemptedCallLogs = Seq[Seq[CallLogs]]
+  // Grouped by FQNS
+  type WorkflowLogs = Map[FullyQualifiedName, AttemptedCallLogs]
+
   implicit class EnhancedFutureFuture[A](val ffa: Future[Future[A]])(implicit ec: ExecutionContext) {
     def flatten: Future[A] = ffa flatMap { fa => fa }
   }